Update data_aggregation
Running the Test: Run the test by navigating to the data_aggregation folder and executing:

python test_data_handler.py

If everything is set up correctly, the test script will print out the price and volume data for Bitcoin. If you see data printed in the console, the module is working!